Add iteration logic to display list
